refactor: member, auth 책임 분리
