+++ api/src/main/java/com/example/muse/domain/member/MemberController.java
@@ -9,23 +9,13 @@ import org.springframework.web.bind.annotation.*;
 public class MemberController {
     private final MemberService memberService;
 
-    @PostMapping("/auth/logout")
-    public T logout() {
-        return memberService.logout();
-    }
-
-    @PostMapping("/auth/reissue")
-    public T reissue() {
-        return memberService.reissue();
-    }
-
-    @GetMapping("/profiles/{memberId}")
-    public T updateProfile(@PathVariable Long memberId) {
-        return memberService.getProfile();
-    }
-
-    @PatchMapping("/profiles/{memberId}")
-    public T updateProfile(@PathVariable Long memberId) {
-        return memberService.updateProfile();
-    }
+//    @GetMapping("/profiles/{memberId}")
+//    public T updateProfile(@PathVariable Long memberId) {
+//        return memberService.getProfile();
+//    }
+//
+//    @PatchMapping("/profiles/{memberId}")
+//    public T updateProfile(@PathVariable Long memberId) {
+//        return memberService.updateProfile();
+//    }
 }
